+++ main/src/main/java/net/oneandone/stool/stage/Stage.java
@@ -196,7 +196,7 @@ public abstract class Stage {
         return file;
     }
 
-    public LocalDateTime birthdate() throws IOException {
+    public LocalDateTime created() throws IOException {
         return LocalDateTime.ofInstant(Instant.ofEpochMilli(creatorFile().getLastModified()), ZoneId.systemDefault());
     }